Refactor code, initialize new restapi app
