--- backend/backend/forms.py
@@ -1,4 +1,0 @@
-from django import forms
-
-class SkeletonForms(forms.Form):
-    test_text = forms.CharField()
