--- backend/backend/templates/backend/skeleton.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <form action="/backend/skeleton-post" method="POST">
-      {% csrf_token %}
-      <input type="text" id="test_text" name="test_text" /><br />
-      <input type="submit" value="OK" />
-    </form>
-    {% if content %}
-    <ul>
-      {% for item in content %}
-
-      <li>{{ item.content }}</li>
-      {% endfor %}
-    </ul>
-
-    {% else %}
-    <p>No content available</p>
-    {% endif %}
-  </body>
-</html>
