+++ backend/requirements.txt
@@ -4,11 +4,5 @@ django-cors-headers==3.13.0
 djangorestframework==3.14.0
 pytz==2022.7.1
 sqlparse==0.4.3
-asgiref==3.6.0
-Django==4.1.5
-django-cors-headers==3.13.0
-djangorestframework==3.14.0
 psycopg2-binary==2.9.5
-pytz==2022.7.1
-sqlparse==0.4.3
 python-dotenv
