+++ backend/restapi/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class BackendConfig(AppConfig):
+class RestapiConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'backend'
+    name = 'restapi'
