+++ frontend/src/Services/ConsultService.js
@@ -1,14 +1,14 @@
 import axios from "axios"
-const baseUrl = "http://127.0.0.1:8000/api/techname/"
+const baseUrl = "http://127.0.0.1:8000/api/consult/"
 
-const getAllTechs = () => {
+const getAllConsults = () => {
   const request = axios.get(baseUrl)
   return request.then((response) => response.data)
 }
 
-const createTech = (newObject) => {
+const createConsult = (newObject) => {
   const request = axios.post(baseUrl, newObject)
   return request.then((response) => response.data)
 }
 
-export default { getAllTechs, createTech }
+export default { getAllConsults, createConsult }
